# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProjectOps API",
    "description": "Autonomous multi-agent project execution and risk management system. Deterministic scheduling, immutable versioned plans, and append-only audits.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/schedule": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "Get Schedule",
        "description": "Run deterministic CPM schedule on active or versioned plan.",
        "operationId": "get_schedule_api_v1_projects__project_id__schedule_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific plan version; defaults to active",
              "title": "Version"
            },
            "description": "Specific plan version; defaults to active"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Override project start date",
              "title": "Start Date"
            },
            "description": "Override project start date"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/schedule/simulate": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "Simulate Schedule",
        "description": "Run what-if scenario simulation without saving to database.\n\nCore Principle 3: Replans are simulated by the deterministic scheduler.",
        "operationId": "simulate_schedule_api_v1_projects__project_id__schedule_simulate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleSimulationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleSimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DependencySchema": {
        "properties": {
          "predecessor_id": {
            "type": "string",
            "title": "Predecessor Id"
          },
          "successor_id": {
            "type": "string",
            "title": "Successor Id"
          },
          "dep_type": {
            "$ref": "#/components/schemas/DependencyType",
            "default": "FINISH_TO_START"
          },
          "lag_days": {
            "type": "integer",
            "title": "Lag Days",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "predecessor_id",
          "successor_id"
        ],
        "title": "DependencySchema",
        "description": "Precedence dependency relationship between tasks."
      },
      "DependencyType": {
        "type": "string",
        "enum": [
          "FINISH_TO_START",
          "START_TO_START",
          "FINISH_TO_FINISH",
          "START_TO_FINISH"
        ],
        "title": "DependencyType",
        "description": "Dependency precedence types (standard project network types)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberWorkloadResponse": {
        "properties": {
          "member_id": {
            "type": "string",
            "title": "Member Id"
          },
          "capacity_hours_per_day": {
            "type": "number",
            "title": "Capacity Hours Per Day"
          },
          "total_assigned_hours": {
            "type": "number",
            "title": "Total Assigned Hours"
          },
          "daily_allocated_hours": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Daily Allocated Hours"
          },
          "peak_daily_hours": {
            "type": "number",
            "title": "Peak Daily Hours"
          },
          "is_overallocated": {
            "type": "boolean",
            "title": "Is Overallocated"
          },
          "overallocated_dates": {
            "items": {
              "type": "string",
              "format": "date"
            },
            "type": "array",
            "title": "Overallocated Dates"
          }
        },
        "type": "object",
        "required": [
          "member_id",
          "capacity_hours_per_day",
          "total_assigned_hours",
          "daily_allocated_hours",
          "peak_daily_hours",
          "is_overallocated",
          "overallocated_dates"
        ],
        "title": "MemberWorkloadResponse",
        "description": "Workload utilization and capacity analysis for a member."
      },
...
      "ScheduleResponse": {
        "properties": {
          "project_start_date": {
            "type": "string",
            "format": "date",
            "title": "Project Start Date"
          },
          "project_finish_date": {
            "type": "string",
            "format": "date",
            "title": "Project Finish Date"
          },
          "target_completion_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Completion Date"
          },
          "scheduled_tasks": {
            "items": {
              "$ref": "#/components/schemas/ScheduledTaskResponse"
            },
            "type": "array",
            "title": "Scheduled Tasks"
          },
          "critical_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Critical Path"
          },
          "member_workloads": {
            "items": {
              "$ref": "#/components/schemas/MemberWorkloadResponse"
            },
            "type": "array",
            "title": "Member Workloads"
          },
          "is_feasible": {
            "type": "boolean",
            "title": "Is Feasible"
          }
        },
        "type": "object",
        "required": [
          "project_start_date",
          "project_finish_date",
          "target_completion_date",
          "scheduled_tasks",
          "critical_path",
          "member_workloads",
          "is_feasible"
        ],
        "title": "ScheduleResponse",
        "description": "Complete computed project schedule output."
      },
      "ScheduleSimulationRequest": {
        "properties": {
          "candidate_tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskCreateSchema"
            },
            "type": "array",
            "title": "Candidate Tasks"
          },
          "candidate_dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencySchema"
            },
            "type": "array",
            "title": "Candidate Dependencies"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "target_completion_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Completion Date"
          }
        },
        "type": "object",
        "title": "ScheduleSimulationRequest",
        "description": "What-if simulation payload (computes impact without writing to DB)."
      },
      "ScheduleSimulationResponse": {
        "properties": {
          "baseline_finish_date": {
            "type": "string",
            "format": "date",
            "title": "Baseline Finish Date"
          },
          "simulated_finish_date": {
            "type": "string",
            "format": "date",
            "title": "Simulated Finish Date"
          },
          "finish_date_delta_working_days": {
            "type": "integer",
            "title": "Finish Date Delta Working Days"
          },
          "critical_path_changed": {
            "type": "boolean",
            "title": "Critical Path Changed"
          },
          "simulation": {
            "$ref": "#/components/schemas/ScheduleResponse"
          }
        },
        "type": "object",
        "required": [
          "baseline_finish_date",
          "simulated_finish_date",
          "finish_date_delta_working_days",
          "critical_path_changed",
          "simulation"
        ],
        "title": "ScheduleSimulationResponse",
        "description": "Comparison diff between active baseline and simulated scenario."
      },
      "ScheduledTaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "early_start": {
            "type": "string",
            "format": "date",
            "title": "Early Start"
          },
          "early_finish": {
            "type": "string",
            "format": "date",
            "title": "Early Finish"
          },
          "late_start": {
            "type": "string",
            "format": "date",
            "title": "Late Start"
          },
          "late_finish": {
            "type": "string",
            "format": "date",
            "title": "Late Finish"
          },
          "duration_days": {
            "type": "integer",
            "title": "Duration Days"
          },
          "total_float": {
            "type": "integer",
            "title": "Total Float"
          },
          "is_critical": {
            "type": "boolean",
            "title": "Is Critical"
          },
          "is_overdue": {
            "type": "boolean",
            "title": "Is Overdue"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "early_start",
          "early_finish",
          "late_start",
          "late_finish",
          "duration_days",
          "total_float",
          "is_critical",
          "is_overdue"
        ],
        "title": "ScheduledTaskResponse",
        "description": "Timing and float metrics for a scheduled task."
      },
      "TaskCreateSchema": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "TODO"
          },
          "estimated_hours": {
            "type": "number",
            "minimum": 0.0,
            "title": "Estimated Hours",
            "default": 8.0
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "TaskCreateSchema",
        "description": "Schema for defining a task when creating or updating a plan snapshot."
      },
...
      "TaskStatus": {
        "type": "string",
        "enum": [
          "TODO",
          "IN_PROGRESS",
          "BLOCKED",
          "COMPLETED"
        ],
        "title": "TaskStatus",
        "description": "Valid task statuses.\nRule: OVERDUE is strictly a derived property, never a stored status."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}